mise en place du dto pour plus de performance

+++ back-end/src/main/java/com/projetMedecine/Modele/Salle.java
@@ -18,7 +18,7 @@ public class Salle {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id_salle")
-    private Integer idSalle;
+    private Long idSalle;
     @Column(name="numero_salle")
     private String numeroSalle;
     @Column(name="nom_salle")
